+++ packages/vite-plugin-cloudflare/src/plugin-config.ts
@@ -3,7 +3,7 @@ import * as path from "node:path";
 import { parseStaticRouting } from "@cloudflare/workers-shared/utils/configuration/parseStaticRouting";
 import * as vite from "vite";
 import { getWorkerConfigs } from "./deploy-config";
-import { hasNodeJsCompat, NodeJsCompat } from "./plugins/nodejs-compat";
+import { hasNodeJsCompat, NodeJsCompat } from "./nodejs-compat";
 import {
 	getValidatedWranglerConfigPath,
 	getWorkerConfig,
@@ -41,6 +41,7 @@ export interface PluginConfig extends EntryWorkerConfig {
 	auxiliaryWorkers?: AuxiliaryWorkerConfig[];
 	persistState?: PersistState;
 	inspectorPort?: number | false;
+	remoteBindings?: boolean;
 	experimental?: Experimental;
 }
 
@@ -58,6 +59,7 @@ interface BaseResolvedConfig {
 	persistState: PersistState;
 	inspectorPort: number | false | undefined;
 	experimental: Experimental;
+	remoteBindings: boolean;
 }
 
 export interface AssetsOnlyResolvedConfig extends BaseResolvedConfig {
@@ -123,6 +125,7 @@ export function resolvePluginConfig(
 	if (viteEnv.isPreview) {
 		return {
 			...shared,
+			remoteBindings: pluginConfig.remoteBindings ?? true,
 			type: "preview",
 			workers: getWorkerConfigs(root),
 		};
@@ -151,6 +154,7 @@ export function resolvePluginConfig(
 			cloudflareEnv,
 			config: entryWorkerResolvedConfig.config,
 			configPaths,
+			remoteBindings: pluginConfig.remoteBindings ?? true,
 			rawConfigs: {
 				entryWorker: entryWorkerResolvedConfig,
 			},
@@ -231,29 +235,10 @@ export function resolvePluginConfig(
 		entryWorkerEnvironmentName,
 		nodeJsCompatMap,
 		staticRouting,
+		remoteBindings: pluginConfig.remoteBindings ?? true,
 		rawConfigs: {
 			entryWorker: entryWorkerResolvedConfig,
 			auxiliaryWorkers: auxiliaryWorkersResolvedConfigs,
 		},
 	};
-}
-
-export function assertIsNotPreview(
-	resolvedPluginConfig: ResolvedPluginConfig
-): asserts resolvedPluginConfig is
-	| AssetsOnlyResolvedConfig
-	| WorkersResolvedConfig {
-	assert(
-		resolvedPluginConfig.type !== "preview",
-		`Expected "assets-only" or "workers" plugin config`
-	);
-}
-
-export function assertIsPreview(
-	resolvedPluginConfig: ResolvedPluginConfig
-): asserts resolvedPluginConfig is PreviewResolvedConfig {
-	assert(
-		resolvedPluginConfig.type === "preview",
-		`Expected "preview" plugin config`
-	);
 }