+++ packages/vite-plugin-cloudflare/src/plugins/additional-modules.ts
@@ -3,42 +3,7 @@ import * as fsp from "node:fs/promises";
 import * as path from "node:path";
 import MagicString from "magic-string";
 import * as vite from "vite";
-import { cleanUrl } from "../utils";
-import { createPlugin } from "./utils";
-
-const ADDITIONAL_MODULE_TYPES = ["CompiledWasm", "Data", "Text"] as const;
-type AdditionalModuleType = (typeof ADDITIONAL_MODULE_TYPES)[number];
-
-const ADDITIONAL_MODULE_PATTERN = `__CLOUDFLARE_MODULE__(${ADDITIONAL_MODULE_TYPES.join("|")})__(.*?)__CLOUDFLARE_MODULE__`;
-export const additionalModuleRE = new RegExp(ADDITIONAL_MODULE_PATTERN);
-const additionalModuleGlobalRE = new RegExp(ADDITIONAL_MODULE_PATTERN, "g");
-
-type ModuleRules = Array<{
-	type: AdditionalModuleType;
-	extensions: string[];
-}>;
-
-const moduleRules: ModuleRules = [
-	{ type: "CompiledWasm", extensions: [".wasm", ".wasm?module"] },
-	{ type: "Data", extensions: [".bin"] },
-	{ type: "Text", extensions: [".txt", ".html"] },
-];
-
-function matchAdditionalModule(source: string) {
-	for (const rule of moduleRules) {
-		for (const extension of rule.extensions) {
-			if (source.endsWith(extension)) {
-				return rule.type;
-			}
-		}
-	}
-
-	return null;
-}
-
-function createModuleReference(type: AdditionalModuleType, id: string) {
-	return `__CLOUDFLARE_MODULE__${type}__${id}__CLOUDFLARE_MODULE__`;
-}
+import { cleanUrl, createPlugin } from "../utils";
 
 /**
  * Plugin to support additional module types (`CompiledWasm`, `Data` and `Text`)
@@ -145,3 +110,37 @@ export const additionalModulesPlugin = createPlugin(
 		};
 	}
 );
+
+const ADDITIONAL_MODULE_TYPES = ["CompiledWasm", "Data", "Text"] as const;
+type AdditionalModuleType = (typeof ADDITIONAL_MODULE_TYPES)[number];
+
+const ADDITIONAL_MODULE_PATTERN = `__CLOUDFLARE_MODULE__(${ADDITIONAL_MODULE_TYPES.join("|")})__(.*?)__CLOUDFLARE_MODULE__`;
+export const additionalModuleRE = new RegExp(ADDITIONAL_MODULE_PATTERN);
+const additionalModuleGlobalRE = new RegExp(ADDITIONAL_MODULE_PATTERN, "g");
+
+type ModuleRules = Array<{
+	type: AdditionalModuleType;
+	extensions: string[];
+}>;
+
+const moduleRules: ModuleRules = [
+	{ type: "CompiledWasm", extensions: [".wasm", ".wasm?module"] },
+	{ type: "Data", extensions: [".bin"] },
+	{ type: "Text", extensions: [".txt", ".html"] },
+];
+
+function matchAdditionalModule(source: string) {
+	for (const rule of moduleRules) {
+		for (const extension of rule.extensions) {
+			if (source.endsWith(extension)) {
+				return rule.type;
+			}
+		}
+	}
+
+	return null;
+}
+
+function createModuleReference(type: AdditionalModuleType, id: string) {
+	return `__CLOUDFLARE_MODULE__${type}__${id}__CLOUDFLARE_MODULE__`;
+}