+++ packages/wrangler/src/cli.ts
@@ -74,4 +74,8 @@ export {
 	convertConfigBindingsToStartWorkerBindings as unstable_convertConfigBindingsToStartWorkerBindings,
 } from "./api";
 
+export { getDetailsForAutoConfig as experimental_getDetailsForAutoConfig } from "./autoconfig/details";
+export { runAutoConfig as experimental_runAutoConfig } from "./autoconfig/run";
+export { Framework as experimental_AutoConfigFramework } from "./autoconfig/frameworks/index";
+
 export { experimental_getWranglerCommands } from "./experimental-commands-api";